Trying to learn Piston
With varying degrees of success I've started working out how to program
Rust and learn Piston at the same time.

For this implementation of Pong I'm working with the Piston event loop
which fires off events of a couple different types. Render, Update, and
then Input. Based on which event gets received, a different handler
deals with the event and changes the underlying game state.

Since I'm not quite clear on how all of the Rust borrowing and such
works the current implementation is naive and I don't have an AI or
bouncing off paddles implemented. My next goal is to figure out how to
have the game state passed to sub functions rather than having every
lambda declared in the main thread.